responses refactored

+++ src/responses/Response.hpp
@@ -21,16 +21,16 @@ public:
     Response(const Request& req, const ServerConfig& config, char *buffer, int buffer_size);
 	Response& operator=(const Response& other);
 
-    void		initializeHttpErrors();
-    void		setStatus(int code);
-	int			getStatusCode();
-    void		addHeader(const std::string& key, const std::string& value);
+    void			initializeHttpErrors();
+    void			setStatus(int code);
+	int				getStatusCode();
+    void			addHeader(const std::string& key, const std::string& value);
 
-	void		generateResponse(const std::string& filename);
-	void		generateDirectoryListing(const std::string& directoryPath);
-	void		generateCGIResponse(const std::string &cgi_response);
-	const char*	getContent();
-	int			getContentLength();
+	void			generateResponse(const std::string& filename);
+	void			generateDirectoryListing(const std::string& directoryPath);
+	void			generateCGIResponse(const std::string &cgi_response);
+	const char*		getContent();
+	int				getContentLength();
 	// const char* toCString();
 
 	/* Exceptions */
@@ -59,8 +59,8 @@ private:
     const char*							_content;
 
 	char*								_buffer;
-	int									_buffer_size;
-	int									_content_length;
+	size_t								_buffer_size;
+	ssize_t								_content_length;
 
     // std::string _readFile(const std::string& filename);
     std::string		_getMimeType(const std::string& filename);
